+++ lib/knossos/algorithm.rb
@@ -1,3 +1,4 @@
+require_relative 'algorithm/aldous_broder'
 require_relative 'algorithm/binary_tree'
 
 module Knossos::Algorithm