+++ spec/knossos/algorithm/aldous_broder_spec.rb
@@ -1,0 +1,5 @@
+require 'knossos/algorithm/shared_examples_for_algorithm'
+
+RSpec.describe Algorithm::AldousBroder do
+  it_behaves_like 'algorithm'
+end